Fix negative turnsToNewPopulation while razing
Now when razing, city's food storage is carried over from the stage when city's population was +1 larger and FoodToNextPopulation was respectively larger. If a player would like to stop razing, such city would spring up immediately growing next turn or further on. Another side effect is negative turnsToNewPopulation in the city screen.

The same surplus cut off could be done to production if a Great Engineer speeds wonder production the same way and if negative turnsToComplete is possible (but have not spotted this line yet).