Fixed a crash when a plane tried to enter a full city (#6652)
This would only occur if the plane was damaged and looking for a spot to 
heal, and had no tile to go to with >0 healing. It would then try to 
head towards the nearest city, regardless of whether it could enter 
that. As it's an air unit, 'head towards' means 'fly there directly this 
turn', which obviously caused crashes if the unit couldn't enter that 
tile.